pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-gem-bot"
version = "0.1.0"
description = "Daily autonomous stock screener with Gemini summary"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "yfinance>=0.2.61",
  "google-genai>=1.7.0",
]

[project.scripts]
stock-gem-bot = "stock_gem_bot.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
